requirements.txt
torch>=2.0.0
torchvision
torchaudio
diffusers>=0.24.0
transformers>=4.35.0
accelerate
safetensors
opencv-python
pillow
numpy>=1.24.0
scikit-image
scikit-learn
fastapi
uvicorn
celery
redis
groundingdino-py
segment-anything
einops
omegaconf
huggingface-hub
tqdm
python-dotenv
pydantic
peft>=0.7.0
